requirements.txt
django==1.5.5
gunicorn==18.0
# Utility Requirements
# for SECURE_KEY generation
lockfile>=0.9
# Uncoment for support ldap
#django-auth-ldap==1.2.0
ansible==2.1.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
butterfly==2.0.2
certifi==2016.9.26
cffi==1.8.3
configobj==4.7.2
cryptography==1.5.2
decorator==3.4.0
enum34==1.1.6
idna==2.1
iniparse==0.4
ipaddress==1.0.17
Jinja2==2.8
MarkupSafe==0.23
meld3==0.6.10
numpy==1.11.2
paramiko==2.0.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.0
pygobject==3.8.2
pygpgme==0.3
pyliblzma==0.5.3
pyOpenSSL==16.1.0
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.12
pyzmq==16.0.0
shadowsocks==2.8.2
singledispatch==3.4.0.3
six==1.10.0
slip==0.4.0
slip.dbus==0.4.0
supervisor==3.1.3
tornado==4.4.2
tornado-systemd==1.0.1
urlgrabber==3.10
websockify==0.6.0
yum-metadata-parser==1.1.4
